gen3/connectivity/roboflow-integration/requirements.txt
depthai~=3.0a
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==9.3.0
blobconverter==1.4.3
roboflow==1.1.36
